+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 ### Fixed
 * Fix some tests that intermittently failed.
+* Fix `isDocumentTranslationComplete()` to reject if document translation fails.
 ### Security